requirements.txt
# conda - python 3.11.0 - Mac M1 Pro
Pillow==9.4.0
opencv-python==4.7.0.72
pandas==2.0.2
confluent-kafka==2.1.1
numpy==1.23.5
PyYAML==6.0
matplotlib==3.7.1
seaborn==0.12.2
jupyterlab==3.5.0
ipython==8.7.0
mediapipe==0.10.0
tensorflow==2.12.0
onnx==1.14.0
onnxruntime==1.15.0
openvino==2023.0.0
weaviate_client==3.19.2
chromadb==0.3.25
snakeviz==2.2.0
pyinstrument==4.5.0